Moved generate_file to helpers
