+++ i18nilize/src/internationalize/helpers.py
@@ -18,3 +18,25 @@ def get_token(file_path):
 def create_json(json_object):
     with open("src/internationalize/jsonFile/translations.json", "w") as outfile:
         outfile.write(json_object)
+
+# Input: None (for now)
+# Output: None, but creates a local JSON file containing translations
+def generate_file():
+    file_content = {
+        "Token": "85124f79-0829-4b80-8b5c-d52700d86e46",
+        "translations" : [{
+				"language": "French",
+				"hello": "bonjour",
+				"No": "Non",
+				"Why": "pourquoi",
+			},
+			{
+				"language": "Spanish",
+				"hello": "Hola",
+			},
+		]
+    }
+
+    # transforms the dictionary object above into a JSON object
+    json_object = json.dumps(file_content, indent=4)
+    create_json(json_object)
